pyproject.toml
[project]
name = "deepvault"
version = "0.1.0"
description = "Enterprise-grade autonomous RAG system for high-performance semantic search and retrieval"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
    "sentence-transformers>=5.3.0",
    "tenacity>=9.1.0",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.15.9",
]

